+++ packages/mongodb-memory-server-core/src/__tests__/MongoMemoryServer.test.ts
@@ -2,6 +2,7 @@
 import { MongoClient } from 'mongodb';
 import * as tmp from 'tmp';
 import MongoMemoryServer, {
+  MongoInstanceData,
   MongoMemoryServerEvents,
   MongoMemoryServerStates,
 } from '../MongoMemoryServer';
@@ -19,7 +20,7 @@ afterEach(() => {
 
 describe('MongoMemoryServer', () => {
   describe('start()', () => {
-    it('should resolve to true if an MongoInstanceData is resolved by _startUpInstance', async () => {
+    it('should not error if an MongoInstanceData is resolved by _startUpInstance', async () => {
       const mongoServer = new MongoMemoryServer();
       jest
         .spyOn(mongoServer, '_startUpInstance')
@@ -28,7 +29,7 @@ describe('MongoMemoryServer', () => {
 
       expect(mongoServer._startUpInstance).not.toHaveBeenCalled();
 
-      await expect(mongoServer.start()).resolves.toEqual(true);
+      await mongoServer.start();
 
       expect(mongoServer._startUpInstance).toHaveBeenCalledTimes(1);
     });
@@ -295,9 +296,11 @@ describe('MongoMemoryServer', () => {
       const mongoServer2 = new MongoMemoryServer();
       // this case can normally happen if "start" is called again
       // @ts-expect-error because "_instanceInfo" is protected
-      mongoServer2._instanceInfo = { instance: { childProcess: {} } } as any; // artificially set this to {} to not be undefined anymore
+      mongoServer2._instanceInfo = {
+        instance: { mongodProcess: {} },
+      } as Partial<MongoInstanceData>; // artificially set this to {} to not be undefined anymore
       await expect(mongoServer2.start()).rejects.toThrow(
-        'Cannot start because "instance.childProcess" is already defined!'
+        'Cannot start because "instance.mongodProcess" is already defined!'
       );
     });
   });
@@ -305,7 +308,7 @@ describe('MongoMemoryServer', () => {
   describe('ensureInstance()', () => {
     it('should throw an error if no "instanceInfo" is defined after calling start', async () => {
       const mongoServer = new MongoMemoryServer();
-      jest.spyOn(mongoServer, 'start').mockImplementationOnce(() => Promise.resolve(true));
+      jest.spyOn(mongoServer, 'start').mockResolvedValue(void 0);
 
       await expect(mongoServer.ensureInstance()).rejects.toThrow(
         'Ensure-Instance failed to start an instance!'
@@ -422,9 +425,7 @@ describe('MongoMemoryServer', () => {
 
   describe('create()', () => {
     it('should create an instance and call ".start"', async () => {
-      jest
-        .spyOn(MongoMemoryServer.prototype, 'start')
-        .mockImplementationOnce(() => Promise.resolve(true));
+      jest.spyOn(MongoMemoryServer.prototype, 'start').mockResolvedValue(void 0);
 
       await MongoMemoryServer.create();